pyproject.toml
[tool.poetry]
name = "course-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "course_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
ipython = "^8.17.2"
pillow = "^10.1.0"
celery = "^5.3.5"
redis = "^5.0.1"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
coverage = "^7.3.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
